+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [0.3.0] - 30 november 2020
 - add support for country code e.g. en-GB
+- remove the need for supportedLanguageCodes (we expect assets for all supportedLocales)
+- simplify YamlLocalizationsDelegate constructor
 
 ## [0.2.0] - 31 october 2020
 - make constructor parameters @required